+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix [#219](https://github.com/shinolab/autd3/issues/219): [C++] inconsistent name autd3::gain::holo::AmplitudeConstraint
 - Fix [#220](https://github.com/shinolab/autd3/issues/220): [C#] missing WithIntensity in Modulation.Static
 - Fix [#221](https://github.com/shinolab/autd3/issues/221): [C++] inconsistent api name Focus::new_with_sampling_configuration
+- Fix [#222](https://github.com/shinolab/autd3/issues/222): [C++] inconsistent api name ModulationWithFreqDiv::with_sampling_configuration
 
 # 17.0.1